finished typing the main bot, fixing some string concat in review, and doing some more formatting
