+++ Anthony_Fantanbot.py
@@ -2,26 +2,19 @@ import logging
 import json
 from functools import wraps
 from telegram import (
-	Update,
-  ReplyKeyboardMarkup,
-  ReplyKeyboardRemove
+	Update
 )
 from telegram.ext import (
 	ApplicationBuilder,
 	ContextTypes,
 	CommandHandler,
-	Updater,
-	MessageHandler,
-	Filters,
-	ConversationHandler,
 	CallbackContext,
-	JobQueue
+	JobQueue,
+	filters
 )
-import schedule
-from threading import Thread
-from time import sleep
 import os
 import datetime
+import pytz
 
 __location__ = os.path.realpath(
 	os.path.join(os.getcwd(), os.path.dirname(__file__)))
@@ -32,8 +25,8 @@ logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s
 
 logger = logging.getLogger(__name__)
 
-vars = json.load(open(os.path.join(__location__,'Telegram.json'),'r'))
-LIST_OF_ADMINS = vars["ADMINS"]
+vars = json.load(open(os.path.join(__location__,'telegram.json'),'r'))
+LIST_OF_ADMINS = vars["admins"]
 
 #########################
 #   Telegram Commands   #
@@ -44,43 +37,53 @@ def restricted(func):
 	Decorator function to only allow admins to send certain commands
 	"""
 	@wraps(func)
-	def wrapped(update,context, *args, **kwargs):
+	def wrapped(update: Update, context: ContextTypes.DEFAULT_TYPE, *args, **kwargs):
 		user_id = update.effective_user.id
 		if user_id not in LIST_OF_ADMINS:
 			update.message.reply_text("ACCESS DENIED!")
 			return
-		return func(update,context, *args, **kwargs)
+		return func(update, context, *args, **kwargs)
 	return wrapped
 
 @restricted
-def CustomeQuery(update,context):
+def CustomeQuery(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 	pass
 
 @restricted
-def GetUserID(update,context):
+async def GetUserID(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 	"""
 	Restricted. Sends the current users id
 	"""
+	user_name = update.effective_user.name
 	user_id = update.effective_user.id
-	update.message.reply_text(str(user_id) + ": Your user ID")
+	await context.bot.send_message(chat_id=vars['debugChatId'], text="User \"{0}\": {1}".format(user_name, user_id))
 
-def SendThicc(update,context):
-	PhotoSender(update, context, 'Media\\THICC.mp4')
+@restricted
+async def GetGroupID(update: Update,context: ContextTypes.DEFAULT_TYPE) -> None:
+	"""
+	Restricted. Sends the current users id
+	"""
+	chat_name = update.effective_chat.effective_name
+	chat_id = update.effective_chat.id
+	await context.bot.send_message(chat_id=vars['debugChatId'], text="Group \"{0}\": {1}".format(chat_name, chat_id))
 
-def SendNut(update,context):
-	PhotoSender(update, context, 'Media\\Nutt.mp4')
+async def SendThicc(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+	await PhotoSender(update, context, 'Media\\THICC.mp4')
 
-def SendJuicy(update,context):
-	PhotoSender(update, context, 'Media\\Grapefruit.mp4')
+async def SendNut(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+	await PhotoSender(update, context, 'Media\\Nutt.mp4')
 
-def PhotoSender(update, context, imagePath: str) -> None:
+async def SendJuicy(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
+	await PhotoSender(update, context, 'Media\\Grapefruit.mp4')
+
+async def PhotoSender(update: Update, context: ContextTypes.DEFAULT_TYPE, imagePath: str) -> None:
 	"""
 	Helper function to send a photo by path
 	"""
 	photo = open(file=os.path.join(__location__,imagePath),mode='rb')
-	context.bot.sendAnimation(chat_id=update.message.chat_id, animation=photo)
+	await context.bot.sendAnimation(chat_id=update.message.chat_id, animation=photo)
 
-def ParseQueue(update, context):
+async def ParseQueue(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 	"""
 	Function to parse the queue. Should move this to another class or db logic
 	"""
@@ -125,7 +128,7 @@ def ParseQueue(update, context):
 	context.bot.send_message(chat_id=vars["SELFID"], text=msgText, parse_mode='Markdown')
 
 @restricted
-def SendReview(update,context):
+async def SendReview(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 	MusicChatID = vars["CHATID"]
 	#MusicChatID = vars["SELFID"]
 	reviewJson = GetNextInQueue()
@@ -143,24 +146,24 @@ def SendReview(update,context):
 	idText = "#AlbumReview No. {0}".format(count + 1)
 	msgBody = "{0}\n\n*Album Title*\n{1}\n\n*Album Artist*\n{2}\n\n*Genre*\n{3}\n\n*Thoughts*\n{4}\n\n*Track Ratings*\n{5}\n\n*Overall Rating*\n{6}\n\n{7}".format(idText,reviewJson["Title"],reviewJson["Artist"],genreTxt,reviewJson["ReviewBody"],TrackList,rating,reviewJson["NextUpText"])
 
-	context.bot.sendPhoto(chat_id=MusicChatID, photo=photo)
-	context.bot.send_message(chat_id=MusicChatID, text=msgBody, parse_mode='Markdown')
+	await context.bot.sendPhoto(chat_id=MusicChatID, photo=photo)
+	await context.bot.send_message(chat_id=MusicChatID, text=msgBody, parse_mode='Markdown')
 
 	UpdateCompleted(reviewJson)
 
-def error(update, context):
+async def error(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
 	"""Log Errors caused by Updates."""
 
-	errorMsg = "Update\n `{0}`\n\n caused error\n `{1}`".format(update,context.error)
+	errorMsg = "Update\n `{0}`\n\n caused error\n `{1}`".format(update, context.error)
 
-	context.bot.send_message(chat_id=vars["SELFID"], text=errorMsg)
+	await context.bot.send_message(chat_id=vars["debugChatId"], text=errorMsg)
 	logger.warning('Update "%s" caused error "%s"', update, context.error) 
 
 #####################
 #   Get Functions   #
 #####################
 
-def GeNextUpText(reviewLst):
+def GeNextUpText(reviewLst: list(json)) -> str:
 	end = False
 	index = 0
 
@@ -178,7 +181,7 @@ def GeNextUpText(reviewLst):
 
 	return NextUpText
     
-def GetNextInQueue():
+def GetNextInQueue() -> json:
 	with open(os.path.join(__location__,'Queue.json'),'r') as file:
 		queue = json.load(file)
 		queueShort = queue["Queue"]
@@ -205,7 +208,7 @@ def GetNextInQueue():
 
 	return review
 
-def GetNumCompleted():
+def GetNumCompleted() -> int:
 	with open(os.path.join(__location__,'Posted_Reviews.json'),'r') as file:
 		completed = json.load(file)
 
@@ -217,7 +220,7 @@ def GetNumCompleted():
 #   Format Functions   #
 ########################
 
-def FormatTracklist(reviewJson):
+def FormatTracklist(reviewJson: json) -> str:
 	trackList = reviewJson["TrackList"]
 
 	tList = ""
@@ -227,17 +230,17 @@ def FormatTracklist(reviewJson):
 
 	return tList
 
-def FormatNextUp(review):
+def FormatNextUp(review: json) -> str:
 	return "NEXT UP:\n\n{0} by {1}\n\n{2}".format(review["Title"],review["Artist"],review["Blurb"])
 
-def FormatRatingBlock(json):
+def FormatRatingBlock(json: json) -> str:
 	txt = "(personal rating + Songs avg) / 2 = Rating\n"
 
 	txt += "( {0} + {1} ) / 2 = {2}".format(json["AlbumFeelingRating"],json["SongAvg"],json["AlbumRating"])
 
 	return txt
 
-def FormatGenreBlock(json):
+def FormatGenreBlock(json: json) -> str:
 	genres = json["Genre"]
 
 	gtxt = ""
@@ -249,7 +252,7 @@ def FormatGenreBlock(json):
 
 	return gtxt
 
-def CheckReviewForValue(queue,equalTo):
+def CheckReviewForValue(queue: json,equalTo: any) -> bool:
 	if queue["Title"] == equalTo:
 		return True
 
@@ -281,7 +284,7 @@ def CheckReviewForValue(queue,equalTo):
 		return True
 	return False
 
-def UpdateCompleted(review):
+def UpdateCompleted(review: json) -> None:
 	with open(os.path.join(__location__,'Posted_Reviews.json'),'r') as file:
 		completed = json.load(file)
 	
@@ -291,45 +294,12 @@ def UpdateCompleted(review):
 	with open(os.path.join(__location__,'Posted_Reviews.json'),'w') as file:
 		json.dump(completed,file,indent=2)
 
-def SendTimedReview(context):
-	#MusicChatID = vars["CHATID"]
-	MusicChatID = vars["SELFID"]
-	reviewJson = GetNextInQueue()
-
-	photo = open(file=reviewJson["AlbumArt"],mode='rb')
-
-	TrackList = FormatTracklist(reviewJson)
-	rating = FormatRatingBlock(reviewJson)
-
-	msgBody = "*Thoughts*\n{0}\n\n*Track Ratings*\n{1}\n\n*Overall Rating*\n{2}".format(reviewJson["ReviewBody"],TrackList,rating)
-
-	context.bot.send_message(chat_id=MusicChatID, text="#AlbumReview")
-	context.bot.sendPhoto(chat_id=MusicChatID, photo=photo)
-	#context.bot.send_message(chat_id=MusicChatID, text=reviewJson["ReviewBody"], parse_mode='Markdown')
-	#context.bot.send_message(chat_id=MusicChatID, text=TrackList)
-	#context.bot.send_message(chat_id=MusicChatID, text=reviewJson["NextUpText"], parse_mode='Markdown')
-	context.bot.send_message(chat_id=MusicChatID, text=msgBody, parse_mode='Markdown')
-	context.bot.send_message(chat_id=MusicChatID, text=reviewJson["NextUpText"], parse_mode='Markdown')
-
-	UpdateCompleted(reviewJson)
-
-def callback_Review(context: CallbackContext):
-	SendTimedReview(context)
-
-def TestSend(update, context):
-	context.bot.send_message(chat_id=vars["SELFID"], text="Henlo", parse_mode='Markdown')
-
-async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
-	await context.bot.send_message(
-		chat_id=update.effective_chat.id,
-		text="I HAVE ARISEN, ONCE MORE!!!"
-	)
-
-def Main():
+def Main() -> None:
 	application = ApplicationBuilder().token(vars['token']).build()
 	job_queue = application.job_queue
 
 	application.add_handler(CommandHandler('getuserid',GetUserID))
+	application.add_handler(CommandHandler('getgroupid',GetGroupID))
 	application.add_handler(CommandHandler('sendreview',SendReview))
 	application.add_handler(CommandHandler('thicc',SendThicc))
 	application.add_handler(CommandHandler('ParseQueue',ParseQueue))
@@ -337,8 +307,10 @@ def Main():
 	application.add_handler(CommandHandler('Juicy',SendJuicy))
 
 	application.add_error_handler(error)
-	tenAm = datetime.datetime(0,0,0,10,0,0)
-	job_weeklyReview = job_queue.run_daily(SendReview,tenAm,[1])
+
+	# Sends the review at 10 am every monday (in theory)
+	tenAm = datetime.time(hour=10, tzinfo=pytz.timezone('America/Chicago'))
+	job_queue.run_daily(SendReview,tenAm,days=(1,1))
 
 	application.run_polling()
 	print('Stopped')
